requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-dotenv==1.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
Pillow==9.0.1
PyJWT==2.3.0
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4
